Refactor to use document helper
